reservation_project/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
diff-match-patch==20241021
Django==5.2.3
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-environ==0.12.0
django-import-export==4.3.7
django-recaptcha==4.1.0
dj-database-url==2.1.0
dnspython==2.7.0
et_xmlfile==2.0.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
mercadopago==2.3.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
psycopg2-binary==2.9.9
python-dotenv==1.1.0
python-slugify==8.0.4
reportlab==4.4.2
requests==2.32.4
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
